// admin-clients/libs/cpanel/sales/data-access/src/lib/sales/payouts/models/payouts-fields.enum.ts
export enum PayoutsFields {
    code = 'code',
    channel = 'channel',
    event = 'event',
    session = 'session',
    seat = 'seat',
    client = 'client',
    clientId = 'clientId',
    payoutType = 'payoutType',
    iban = 'iban',
    bacs = 'bacs',
    price = 'price',
    purchaseDate = 'purchaseDate',
    status = 'status',
    actions = 'actions'
}